Build six different files
